Some last changes before bed
I wanted to get batch inserts working before bed
but it is definitely time for me to be asleep. It
should work for user but there isn't a way to call
it, I need to add a button or something. I think all
of the 'insert' stored procedures are done but we will
still need to do a whole bunch of updates and things
of the sort. Also got the css working and made it look
slightly nicer. Joyce helped with that part.